- 評価関数ファイルを読み込んで、C++の文字列literalに変換するスクリプト追加。   - ここで変換した.cppをやねうら王のプロジェクトに追加してビルドする。   - ビルドする時に、EVAL_EMBEDDINGをdefineしてビルドする。
- 一つ前のcommitでビルドできなくなっていたの修正。

+++ source/book/makebook2023.cpp
@@ -1262,7 +1262,9 @@ namespace MakeBook2023
 
 					// nの64倍ぐらいで終わるんちゃうんか？
 					progress.reset(n * 64);
-					u64 c = 0;
+
+					// 進捗出力用カウンター
+					size_t c = 0;
 					StateInfo si;
 
 					// カスタム比較関数
